requirements.txt
annotated-types==0.7.0
anyio==4.11.0
cffi==2.0.0
cryptography==46.0.2
ecdsa==0.19.1
fastapi==0.119.0
greenlet==3.2.4
idna==3.11
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
